test_data/requirements.txt
asgiref==3.9.1
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
Faker==37.5.3
idna==3.10
mypy_extensions==1.1.0
neo4j==5.28.2
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
pytz==2025.2
requests==2.32.4
sqlparse==0.5.3
tabulate==0.9.0
tzdata==2025.2
urllib3==2.5.0
